+++ R/load.r
@@ -2,6 +2,7 @@ load = function (..., isolate = FALSE, renv = TRUE, dotenv = TRUE, dev = TRUE) {
   check_dots_empty()
 
   if (renv) {
+    configure_renv()
     load_renv()
     disable_renv_autoloader()
   }