+++ R/renv.r
@@ -1,18 +1,21 @@
 load_renv = function () {
   if (file.exists('renv/activate.R')) {
-    renv_ignored_pkgs = getOption(
-      'renv.settings.ignored.packages',
-      default = getOption('renv.settings', list())$ignored.packages
-    )
-    options(
-      # Have ‘renv’ ignore ‘rprofile’ to avoid polluting the ‘renv’ status.
-      renv.settings.ignored.packages = c(renv_ignored_pkgs, .packageName),
-      # Disable the loading of a global `.Rprofile` file via ‘renv’ (if set by the user) since that conflicts with ‘rprofile’.
-      renv.config.user.profile = FALSE
-    )
-
     sys.source('renv/activate.R', envir = .GlobalEnv)
   }
+}
+
+configure_renv = function () {
+  renv_ignored_pkgs = getOption(
+    'renv.settings.ignored.packages',
+    default = getOption('renv.settings', list())$ignored.packages
+  )
+
+  options(
+    # Have ‘renv’ ignore ‘rprofile’ to avoid polluting the ‘renv’ status.
+    renv.settings.ignored.packages = c(renv_ignored_pkgs, .packageName),
+    # Disable the loading of a global `.Rprofile` file via ‘renv’ (if set by the user) since that conflicts with ‘rprofile’.
+    renv.config.user.profile = FALSE
+  )
 }
 
 disable_renv_autoloader = function () {